+++ lib/mail-iso-2022-jp/patches.rb
@@ -43,11 +43,35 @@ module Mail
     include FieldWithIso2022JpEncoding
   end
 
+  class SenderField < StructuredField
+    include FieldWithIso2022JpEncoding
+  end
+
   class ToField < StructuredField
     include FieldWithIso2022JpEncoding
   end
 
   class CcField < StructuredField
+    include FieldWithIso2022JpEncoding
+  end
+
+  class ReplyToField < StructuredField
+    include FieldWithIso2022JpEncoding
+  end
+
+  class ResentFromField < StructuredField
+    include FieldWithIso2022JpEncoding
+  end
+
+  class ResentSenderField < StructuredField
+    include FieldWithIso2022JpEncoding
+  end
+
+  class ResentToField < StructuredField
+    include FieldWithIso2022JpEncoding
+  end
+
+  class ResentCcField < StructuredField
     include FieldWithIso2022JpEncoding
   end
 end